move logic to AccountData
